Add sample unit test using mocha, express, and expect.
The test checks whether the server sends a valid response for a request
to '/users'. This succeeds in the VM, but will fail at travis currently,
so the test will not run by default just yet. It is integrated with grunt
already, so activating is only a small step away. If `grunt-cli` is
installed globally, it can be run like this: `grunt mochaTest`

I followed a mixture of the following guides:
* https://semaphoreci.com/community/tutorials/getting-started-with-node-js-and-mocha
* http://webapplog.com/test-driven-development-in-node-js-with-mocha/